+++ sources/src/main/java/com/google/solutions/jitaccess/core/activation/JitActivationRequest.java
@@ -22,7 +22,6 @@
 package com.google.solutions.jitaccess.core.activation;
 
 import com.google.solutions.jitaccess.core.UserId;
-import com.google.solutions.jitaccess.core.entitlements.EntitlementId;
 
 import java.time.Instant;
 import java.util.Collection;
