+++ sources/src/main/java/com/google/solutions/jitaccess/core/activation/MpaActivationRequest.java
@@ -23,7 +23,6 @@ package com.google.solutions.jitaccess.core.activation;
 
 import com.google.common.base.Preconditions;
 import com.google.solutions.jitaccess.core.UserId;
-import com.google.solutions.jitaccess.core.entitlements.EntitlementId;
 
 import java.time.Instant;
 import java.util.Collection;
