+++ sources/src/test/java/com/google/solutions/jitaccess/core/activation/TestActivationRequest.java
@@ -21,9 +21,7 @@
 
 package com.google.solutions.jitaccess.core.activation;
 
-import com.google.solutions.jitaccess.core.AccessException;
 import com.google.solutions.jitaccess.core.UserId;
-import com.google.solutions.jitaccess.core.entitlements.EntitlementId;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
